wrapspawner/requirements.txt
dockerspawner==12.0.0
nest-asyncio==1.5.1
rich==10.6.0